services/router/requirements.txt
grpcio==1.78.0
protobuf==6.31.1
certifi==2026.1.4
charset-normalizer==3.4.4
DAWG-Python==0.7.2
docopt==0.6.2
filelock==3.20.3
fsspec==2026.1.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
intervaltree==3.2.1
ipymarkup==0.9.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
natasha==1.6.0
navec==0.10.0
networkx==3.6.1
numpy==2.4.1
packaging==25.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
PyYAML==6.0.3
razdel==0.5.0
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
setuptools==80.9.0
slovnet==0.6.0
sortedcontainers==2.4.0
sympy==1.14.0
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.6
typing_extensions==4.15.0
urllib3==2.6.3
yargy==0.16.0
